requirements.txt
absl-py==2.1.0
accelerate==0.29.2
aiohttp==3.9.3
aiosignal==1.3.1
annotated-types==0.6.0
anyascii==0.3.2
anyio==4.3.0
argbind==0.3.7
asttokens==2.4.1
astunparse==1.6.3
attrs==23.2.0
audioread==3.0.1
Babel==2.14.0
bangla==0.0.2
blinker==1.7.0
blis==0.7.11
bnnumerizer==0.0.2
bnunicodenormalizer==0.1.6
catalogue==2.0.10
certifi==2024.2.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cloudpathlib==0.16.0
confection==0.1.4
contourpy==1.2.1
coqpit==0.0.17
cycler==0.12.1
cymem==2.0.8
Cython==3.0.10
datasets==2.18.0
dateparser==1.1.8
decorator==5.1.1
descript-audio-codec==1.0.0
descript-audiotools==0.7.2
dill==0.3.8
distro==1.9.0
docopt==0.6.2
docstring_parser==0.16
einops==0.7.0
encodec==0.1.1
executing==2.0.1
ffmpy==0.3.2
filelock==3.13.4
fire==0.6.0
Flask==3.0.3
flatbuffers==24.3.25
flatten-dict==0.4.2
fonttools==4.51.0
frozenlist==1.4.1
fsspec==2024.2.0
future==1.0.0
g2pkk==0.1.2
gast==0.5.4
google-pasta==0.2.0
grpcio==1.62.1
gruut==2.2.3
gruut-ipa==0.13.0
gruut-lang-de==2.0.0
gruut-lang-en==2.0.0
gruut-lang-es==2.0.0
gruut-lang-fr==2.0.2
h11==0.14.0
h5py==3.11.0
hangul-romanize==0.1.0
httpcore==1.0.5
httpx==0.27.0
huggingface-hub==0.22.2
idna==3.7
importlib_resources==6.4.0
inflect==7.2.0
ipython==8.23.0
itsdangerous==2.1.2
jamo==0.4.1
jedi==0.19.1
jieba==0.42.1
Jinja2==3.1.3
joblib==1.4.0
jsonlines==1.2.0
julius==0.2.7
keras==3.2.1
kiwisolver==1.4.5
langcodes==3.3.0
lazy_loader==0.4
libclang==18.1.1
librosa==0.10.1
llvmlite==0.42.0
Markdown==3.6
markdown-it-py==3.0.0
markdown2==2.4.13
MarkupSafe==2.1.5
matplotlib==3.8.4
matplotlib-inline==0.1.6
mdurl==0.1.2
ml-dtypes==0.3.2
more-itertools==10.2.0
mpmath==1.3.0
msgpack==1.0.8
multidict==6.0.5
multiprocess==0.70.16
murmurhash==1.0.10
namex==0.0.7
networkx==2.8.8
nltk==3.8.1
num2words==0.5.13
numba==0.59.1
numpy==1.26.4
openai==1.17.0
opt-einsum==3.3.0
optree==0.11.0
packaging==24.0
pandas==1.5.3
parler_tts @ git+https://github.com/huggingface/parler-tts.git@10016fb0300c0dc31a0fb70e26f3affee7b62f16
parso==0.8.4
pexpect==4.9.0
pillow==10.3.0
platformdirs==4.2.0
pooch==1.8.1
preshed==3.0.9
prompt-toolkit==3.0.43
protobuf==4.25.3
psutil==5.9.8
ptyprocess==0.7.0
pure-eval==0.2.2
pyarrow==15.0.2
pyarrow-hotfix==0.6
pycparser==2.22
pydantic==2.7.0
pydantic_core==2.18.1
pydub==0.25.1
Pygments==2.17.2
pyloudnorm==0.1.1
pynndescent==0.5.12
pyparsing==3.1.2
pypinyin==0.51.0
pysbd==0.3.4
pystoi==0.4.1
python-crfsuite==0.9.10
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.1
randomname==0.2.1
regex==2023.12.25
requests==2.31.0
rich==13.7.1
safetensors==0.4.2
scikit-learn==1.4.2
scipy==1.13.0
sentencepiece==0.2.0
simpleaudio==1.0.4
six==1.16.0
smart-open==6.4.0
sniffio==1.3.1
sounddevice==0.4.6
soundfile==0.12.1
soxr==0.3.7
spacy==3.7.4
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
stack-data==0.6.3
SudachiDict-core==20240409
SudachiPy==0.6.8
sympy==1.12
tensorboard==2.16.2
tensorboard-data-server==0.7.2
tensorflow==2.16.1
tensorflow-io-gcs-filesystem==0.36.0
termcolor==2.4.0
tf_keras==2.16.0
thinc==8.2.3
threadpoolctl==3.4.0
tokenizers==0.15.2
torch==2.2.2
torch-stoi==0.2.1
torchaudio==2.2.2
tqdm==4.66.2
trainer==0.0.36
traitlets==5.14.2
-e git+https://github.com/huggingface/transformers.git@0bd58f1ce0573c0e3269de4215a17d318add49b9#egg=transformers
TTS==0.22.0
typeguard==4.2.1
typer==0.9.4
typing_extensions==4.11.0
tzdata==2024.1
tzlocal==5.2
umap-learn==0.5.6
Unidecode==1.3.8
urllib3==2.2.1
wasabi==1.1.2
wcwidth==0.2.13
weasel==0.3.4
Werkzeug==3.0.2
wrapt==1.16.0
xxhash==3.4.1
yarl==1.9.4
